services/slb.go:SoftwareLoadBalancer, Slb, New
This adds the Software LoadBlancer, and its initiation with the New function:
The SoftwareLoadBalancer can be used as a handler with http.Server.
Please note that this is not tested at this stage, as the requirement is to run this with the not yet implemented Run() method.

balance/services/slb_test/go:added series of tests for configuration

balance/pkg/mock/mock.go:Added mocks for tests

Signed-off-by: royalt-lb <royalt@lightbitslabs.com>