services/slb/selector.go:Added Endpoints() to inteface
Added Endpoints() to interface to retrieve the current Endpoints
list, as there was no way in retrieving this before.

This is imporant for implementations where a user wants to verify whether
an enpoint exists before adding/removing the endpoint, or to compare to the
configuration provided originally, hence it is returned as an array, the same
as in the configuration.

When implementing this, please ensure that if endpoints are contained in the original
configuration, that these would remain in the same order in the output of Endpoints().

Signed-off-by: royalt-lb <royalt@lightbitslabs.com>