services/slb/selectors/roundRobin:Fix omit first & last endpoint
The selector of the round robin nlb would change the index before accessing the endpoint list, which would cause
it to skip the first endpoint.
Due to the above the index would reach len-1 prematurely, and reset the index to early.

With the fix the index is changed only in defer, which means post accessing the endpoint it needs to access.
It will only be increased if not reset and vice versa.

Signed-off-by: royalt-lb <royalt@lightbitslabs.com>