tests/integration/random_test.go: Random Selector Tests
This adds an integration test for the Random Selector.
Verifying that it does not select sequentially.
This does not validate actual randomness, as this would be difficult to
determine. So technically just returning the reversed list would already
provide the correct result here to fulfill the test requirements, but
that is beyond the scope for this test.

Signed-off-by: royalt-lb <royalt@lightbitslabs.com>