tests/integration:add mTls Compliance
setup:
In order to use mTls the service needs to have certificates
generated. In this constellation the machine running the test is the api client and the
service (slb docker) the host.

As the host cert needs to incorporate the host's IP address the docker needs to run and remain running to not loose it's IP.
Hence copying the certificates to the docker without a running docker, or commiting the container and then starting it were not a valiable solution.

In setup the auth certificates both for the server and the
client are created, and stored where the temproary directory for the test certificates is.
These certificates are seen by the host as the directory is mounted to the hots.

In order for the service not to start running prior to having the
certificates, a waiter script is added to the entrypoint of the
container.

The above required a behavior in the way tests containers are named as
well.
Naming is now uniform with test.name.uuid-containertype

Signed-off-by: royalt-lb <royalt@lightbitslabs.com>