Dockerfile.balance-build:cache unit tests
Running unit tests in a docker image is not cached,
hence every build for container creatiion caused a rebuild of the balance container for each run.
This caused a lot unnessecary image builds,
and overhead of builds in terms of build time and resources.

If the output is cached, this causes the cmd output to be cached within the docker build.
This dramatically improved container creation time.

Signed-off-by: royalt-lb <royalt@lightbitslabs.com>