test/integration/docker_deployment.go:wrap harness as deployment
This extracts the docker deployment setup to be a TestDeployment, and extracts the harness so it could be used in system tests as well.
This streamlines the code and separates deployment more clearly from
tests and test utitlities, as it
is intended to allow for simple replacement by different deployment
types.

Signed-off-by: roy4ik <royalt@lightbitslabs.com>