slb/slb_test.go:stabilize happu flow
Although locally this barely ever occured, this test was flaky when
running in the CI or a containerized environment, erroring out on a
connection error, due to the slb not having properly started prior to
sending requests.

In order to mitigate this issue a short delay was added before sening
requests.
Additionally, run error checking was improved.

Signed-off-by: roy4ik <royalt@lightbitslabs.com>